refactor: separate examples to new files
